Cleaned up the github repository and deleted some unnecessary files
