+++ ReadData.py
@@ -1,22 +1,31 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Wed Feb 12 23:35:22 2025
-
-@author: makro
-"""
-
 import pandas as pd
 import os
+import fastparquet
 
-# Define the path to the directory containing the data
-DATA_DIR = os.path.join(os.getcwd(), 'Data')
+# Define the absolute path to the directory containing the data
+DATA_DIR = r"C:\Users\danie\OneDrive - Delft University of Technology\Q3\Machine Learning\DATA_TRAINING_GREEN"
 
 # Initialize a dictionary to store data
 nyc_taxi_data = {}
 
+# Required columns
+columns_to_keep = ["tpep_pickup_datetime", "lpep_pickup_datetime", "RatecodeID", "trip_distance",
+                   "fare_amount", "tip_amount", "payment_type"]
+
 # Loop through all files in the directory
 for file in os.listdir(DATA_DIR):
     if file.endswith(".parquet"):
+        # print(f"Processing file: {file}")
+        # Identify taxi type (Green or Yellow) from filename
+        if "green" in file.lower():
+            taxi_type = "green"
+        elif "yellow" in file.lower():
+            taxi_type = "yellow"
+        else:
+            print(f"Skipping file {file} due to unknown taxi type.")
+            continue
+
+        # print(f"Taxi type for {file}: {taxi_type}")
         # Extract year and month from the filename
         try:
             parts = file.split('_')[-1].split('-')  # Extracting '2024-01.parquet'
@@ -30,10 +39,14 @@ for file in os.listdir(DATA_DIR):
         file_path = os.path.join(DATA_DIR, file)
         df = pd.read_parquet(file_path)
 
+        df = df.loc[:, df.columns.intersection(columns_to_keep)]
+        df["taxi_type"] = taxi_type
+
         # Store the DataFrame in a nested dictionary
         if year not in nyc_taxi_data:
             nyc_taxi_data[year] = {}
         nyc_taxi_data[year][month] = df
+
 
 # Display available years and months
 years = sorted(nyc_taxi_data.keys())
@@ -41,9 +54,24 @@ print(f"Data available for years: {years}")
 for year in years:
     print(f"Year {year}: Months available -> {sorted(nyc_taxi_data[year].keys())}")
 
-# Example usage: Access January 2024 data
-year, month = 2024, 1
-if year in nyc_taxi_data and month in nyc_taxi_data[year]:
-    print(nyc_taxi_data[year][month].head())
+pd.set_option('display.max_columns', None)
 
-    print(nyc_taxi_data[year][month].head())
+# Example usage: Access January 2024 data
+def get_taxi_data(year, month):
+    if year in nyc_taxi_data and month in nyc_taxi_data[year]:
+        return nyc_taxi_data[year][month]
+    else:
+        print(f"No data available for {year}-{month}.")
+        return None
+
+df_jan_2024 = get_taxi_data(2021, 1)
+if df_jan_2024 is not None:
+    print(df_jan_2024.head())
+    # print(df_jan_2024.tail())
+
+
+
+
+
+
+
